+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v2.2.0-dev
+## v2.2.0 (2022-10-16)
 
 - Enhancements
     - Configuration access for a connection now returns the configured `:otp_app` used (`nil` if none)